Initiate reconnect when server stops sending data
If there is a network connectivity issue with the server, we could end
up waiting for a TCP timeout before detecting that our pings are not
being responded to.

Instead, let's keep track of the last time the server sent us a message
and initiate a reconnect if we haven't received anything within a
sessionTimeout interval.