Don't clear connectTimeout until ConnectResponse is received
Otherwise we leave a window between when the TCP connection emits
'connect' and we've sent the ConnectRequest, and when the
ConnectResponse comes back, when we have no timeout applying at
all and no attempt to talk to the socket again.

If we get netsplit from ZK during such a window, we may hang in
the CONNECTING state indefinitely.